refactor: rename loadModule to createModuleLoader
