+++ lib/createModuleLoader.js
